dev-requirements.txt
-r requirements.txt

# dev requirements
pytest~=8.3.2
pytest-cov
pytest-flask
pytest-docker~=3.1.1

docker~=7.1.0

hatchling


